[MOVE] event: reorganize files according to the guidelines.
No change has been made in this commit, only file move and renaming.
