+++ addons/event/__openerp__.py
@@ -23,18 +23,18 @@ Key Features
         'security/event_security.xml',
         'security/ir.model.access.csv',
         'wizard/event_confirm_view.xml',
-        'event_view.xml',
+        'views/event_views.xml',
         'report/report_event_registration_view.xml',
         'report/event_event_templates.xml',
         'report/event_event_reports.xml',
-        'event_data.xml',
-        'event_tip_data.xml',
-        'res_config_view.xml',
-        'views/event.xml',
-        'email_template.xml',
+        'data/event_data.xml',
+        'data/event_data_tip.xml',
+        'views/event_config_settings_views.xml',
+        'views/event_templates.xml',
+        'data/email_template_data.xml',
     ],
     'demo': [
-        'event_demo.xml',
+        'data/event_demo.xml',
     ],
     'installable': True,
     'auto_install': False,
