+++ addons/event/data/email_template_data.xml
