+++ addons/event/data/event_demo.xml
