+++ addons/event/models/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import models
-import wizard
-import report
+import event
+import event_mail
+import event_config_settings
