+++ addons/event/models/event_config_settings.py
