+++ addons/event/models/event_mail.py
