+++ addons/event/views/event_views.xml
